feat: 탐험 해금 영속화 및 게스트 제한 구현 #41
- Exploration DataSource/Repository 구현 (SharedPreferences 영속화)
- 시드 데이터 기반 행성·지역 초기화 및 해금 상태 병합
- 순차 해금 강제: 이전 행성 미클리어 시 다음 행성 잠금
- 게스트 사용자 탐험 진입 시 로그인 유도 다이얼로그
- 공통 해금 다이얼로그 유틸리티(unlock_dialog_helper) 추출
- 게스트 로그아웃 시 탐험·연료·할일·타이머 데이터 클리어

+++ lib/features/exploration/domain/entities/exploration_node_entity.dart
@@ -53,14 +53,14 @@ class ExplorationNodeEntity {
   /// 계층 깊이 (0=galaxy, 1=starSystem, 2=planet, 3=region)
   final int depth;
 
-  /// 아이콘 (이모지: 🌍, 🇰🇷 등)
+  /// 아이콘 (행성: 이모지 🌍 / 지역: 국가 코드 KR)
   final String icon;
 
   /// 상위 노드 ID (null = 최상위)
   final String? parentId;
 
   /// 해금에 필요한 연료
-  final double requiredFuel;
+  final int requiredFuel;
 
   /// 해금 여부
   final bool isUnlocked;
@@ -88,7 +88,7 @@ class ExplorationNodeEntity {
     int? depth,
     String? icon,
     String? parentId,
-    double? requiredFuel,
+    int? requiredFuel,
     bool? isUnlocked,
     bool? isCleared,
     int? sortOrder,
